classifier/python/app/requirements.txt
alembic==1.0.11
asn1crypto==0.24.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bokeh==1.3.4
boto3==1.9.162
botocore==1.12.163
category-encoders==2.0.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
colorama==0.4.1
configparser==3.7.4
croniter==0.3.30
cryptography==2.7
cycler==0.10.0
Cython==0.29.12
cytoolz==0.10.0
dask==2.2.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.0
docutils==0.15.2
entrypoints==0.3
funcsigs==1.0.2
future==0.17.1
geographiclib==1.49
geopy==1.20.0
gitdb2==2.0.5
GitPython==2.1.13
gunicorn==19.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json-merge-patch==0.2
json5==0.8.5
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.5.0
kiwisolver==1.1.0
lockfile==0.12.2
lxml==4.4.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
munch==2.3.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.4
notebook==6.0.0
numpy==1.16.4
olefile==0.46
ordereddict==1.1
packaging==19.0
pandas==0.25.0
pandas-profiling==1.4.1
pandocfilters==1.4.2
parso==0.5.0
patsy==0.5.1
pendulum==2.0.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
psycopg2==2.7.6.1
ptyprocess==0.6.0
py4j==0.10.8.1
pyarrow==0.11.1
pyasn1==0.4.6
pycparser==2.19
Pygments==2.4.2
PyHive==0.6.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyQt5-sip==4.19.18
pyrsistent==0.14.11
PySocks==1.7.0
python-daemon==2.2.3
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.1
pytzdata==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.0.0
rdflib==4.2.2
requests==2.22.0
retrying==1.3.3
rsa==4.0
s3transfer==0.2.1
sasl==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
setproctitle==1.1.10
sip==4.19.8
six==1.12.0
smmap2==2.0.5
soupsieve==1.9.2
SQLAlchemy==1.3.7
sqlalchemy-redshift==0.7.3
statsmodels==0.10.1
tabulate==0.8.3
tenacity==5.0.4
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
thrift==0.11.0
thrift-sasl==0.3.0
toolz==0.10.0
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
typing==3.6.4
tzlocal==2.0.0
ua-parser==0.8.0
unicodecsv==0.14.1
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
WTForms==2.2.1
zope.deprecation==4.4.0
grpcio==1.24.3
grpcio-tools==1.24.3
